Use Jdk heap buffer pooling in Reactive SQL Clients
See #42224

To avoid having several sys props for config, reuse the existing one related to http server.
